708 - Move Natural Key Manager back to models to avoid circular import
